Improve global variable treatment
Stop checking for parent namespaces of the global variable (for example,
for mw.config.get, don't check mw.config too).

When a global identifier is not found to be defined, like mw.config.get,
check if the sub namespaces like mw.config are somewhere as an
assignment, and if they are, don't complain about the identifier not
being defined since a parent namespace is an opaque assignment.