Make failure to walk a JS file non-fatal
Bug: https://phabricator.wikimedia.org/T186284

For some reason the walking of the hogan.js is failing but it is not clear.

In a previous commit the dependencies have been updated in case it was an acorn
issue, and here we'll make the error non-fatal.

Instead of breaking out with the exception, log it to the console including the
stack trace, and let the process continue.