Don't treat multiple definitions of global variables as errors
For example, when defining mw.track in multiple files in JS in core,
that shouldn't be treated as an error.

Instead, try all of the definer files to check for existence on the
dependency tree, and if at least one of them is included without errors,
then do not fail at all.

If all the definer files fail with dependency errors, instead consider
it a "not_found" error, and add all the errors to the logs to show the
user the files that should be in the dependency tree.